Set get_time to ms & add coder process simulation

+++ utils/utils.h
@@ -6,7 +6,7 @@
 /*   By: flinguen <florent@linguenheld.net>          +#+  +:+       +#+       */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/05/15 17:43:34 by flinguen          #+#    #+#             */
-/*   Updated: 2026/05/15 17:43:34 by flinguen         ###   ########.fr       */
+/*   Updated: 2026/05/15 23:11:36 by flinguen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,6 @@
  * @brief Get the current time whit the gettimeofday function.
  * @return time in milliseconds
  */
-int	get_time(void);
+long	get_time(void);
 
 #endif
